Reduce frame images frame index to 32 bit.

+++ common/TracyQueue.hpp
@@ -144,7 +144,7 @@ struct QueueFrameMark
 
 struct QueueFrameImageLean
 {
-    uint64_t frame;
+    uint32_t frame;
     uint16_t w;
     uint16_t h;
     uint8_t flip;
